requirements.txt
pandas
datasets
pyarrow
langchain
langchain-text-splitters
python-dotenv
snowflake-connector-python
snowflake-sqlalchemy